Refactor checkbox and checkbox group components to align with shadcn ui style.
